added files for motor control

+++ leobot_firmware/STM32F4_LeoBot/Libs/rosserial/stm32f4hardware.h
@@ -30,7 +30,7 @@ public:
 	void write(uint8_t* data, uint32_t length);
 
 	// Returns milliseconds since start of program
-	unsigned long time();
+	uint32_t time();
 
 };
